requirements.txt
aiogram==3.*
python_dotenv